+++ torch/quantization/fx/quantization_patterns.py
@@ -6,7 +6,7 @@ If you are adding a new entry/functionality, please, add it to the
 appropriate files under `torch/ao/quantization/fx/`, while adding an import statement
 here.
 """
-from torch.ao.quantization.fx.quantization_patterns import (
+from torch.ao.quantization.fx.quantize_handler import (
     QuantizeHandler,
     BinaryOpQuantizeHandler,
     CatQuantizeHandler,
